// src/discord/generated/PreloadedUserSettings.ts
export interface PreloadedUserSettings_GuildFolder {
    /**
     * @generated from protobuf field: repeated fixed64 guild_ids = 1
     */
    guildIds: bigint[];
    /**
     * @generated from protobuf field: optional google.protobuf.Int64Value id = 2
     */
    id?: Int64Value;
    /**
     * @generated from protobuf field: optional google.protobuf.StringValue name = 3
     */
    name?: StringValue;
    /**
     * @generated from protobuf field: optional google.protobuf.UInt64Value color = 4
     */
    color?: UInt64Value;
}
/**
 * @generated from protobuf message discord_protos.discord_users.v1.PreloadedUserSettings.GuildFolders
 */
export interface PreloadedUserSettings_GuildFolders {
    /**
     * @generated from protobuf field: repeated discord_protos.discord_users.v1.PreloadedUserSettings.GuildFolder folders = 1
     */
    folders: PreloadedUserSettings_GuildFolder[];
    /**
     * @generated from protobuf field: repeated fixed64 guild_positions = 2
     */
    guildPositions: bigint[];
}
// ...
export const PreloadedUserSettings = new PreloadedUserSettings$Type();
// ...
class PreloadedUserSettings_GuildFolder$Type extends MessageType<PreloadedUserSettings_GuildFolder> {
    constructor() {
        super("discord_protos.discord_users.v1.PreloadedUserSettings.GuildFolder", [
            { no: 1, name: "guild_ids", kind: "scalar", repeat: 1 /*RepeatType.PACKED*/, T: 6 /*ScalarType.FIXED64*/, L: 0 /*LongType.BIGINT*/ },
            { no: 2, name: "id", kind: "message", T: () => Int64Value },
            { no: 3, name: "name", kind: "message", T: () => StringValue },
            { no: 4, name: "color", kind: "message", T: () => UInt64Value }
        ]);
    }
    create(value?: PartialMessage<PreloadedUserSettings_GuildFolder>): PreloadedUserSettings_GuildFolder {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.guildIds = [];
        if (value !== undefined)
            reflectionMergePartial<PreloadedUserSettings_GuildFolder>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PreloadedUserSettings_GuildFolder): PreloadedUserSettings_GuildFolder {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated fixed64 guild_ids */ 1:
                    if (wireType === WireType.LengthDelimited)
                        for (let e = reader.int32() + reader.pos; reader.pos < e;)
                            message.guildIds.push(reader.fixed64().toBigInt());
                    else
                        message.guildIds.push(reader.fixed64().toBigInt());
                    break;
                case /* optional google.protobuf.Int64Value id */ 2:
                    message.id = Int64Value.internalBinaryRead(reader, reader.uint32(), options, message.id);
                    break;
                case /* optional google.protobuf.StringValue name */ 3:
                    message.name = StringValue.internalBinaryRead(reader, reader.uint32(), options, message.name);
                    break;
                case /* optional google.protobuf.UInt64Value color */ 4:
                    message.color = UInt64Value.internalBinaryRead(reader, reader.uint32(), options, message.color);
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PreloadedUserSettings_GuildFolder, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated fixed64 guild_ids = 1; */
        if (message.guildIds.length) {
            writer.tag(1, WireType.LengthDelimited).fork();
            for (let i = 0; i < message.guildIds.length; i++)
                writer.fixed64(message.guildIds[i]);
            writer.join();
        }
        /* optional google.protobuf.Int64Value id = 2; */
        if (message.id)
            Int64Value.internalBinaryWrite(message.id, writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* optional google.protobuf.StringValue name = 3; */
        if (message.name)
            StringValue.internalBinaryWrite(message.name, writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* optional google.protobuf.UInt64Value color = 4; */
        if (message.color)
            UInt64Value.internalBinaryWrite(message.color, writer.tag(4, WireType.LengthDelimited).fork(), options).join();
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message discord_protos.discord_users.v1.PreloadedUserSettings.GuildFolder
 */
export const PreloadedUserSettings_GuildFolder = new PreloadedUserSettings_GuildFolder$Type();
// @generated message type with reflection information, may provide speed optimized methods
class PreloadedUserSettings_GuildFolders$Type extends MessageType<PreloadedUserSettings_GuildFolders> {
    constructor() {
        super("discord_protos.discord_users.v1.PreloadedUserSettings.GuildFolders", [
            { no: 1, name: "folders", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => PreloadedUserSettings_GuildFolder },
            { no: 2, name: "guild_positions", kind: "scalar", repeat: 1 /*RepeatType.PACKED*/, T: 6 /*ScalarType.FIXED64*/, L: 0 /*LongType.BIGINT*/ }
        ]);
    }
    create(value?: PartialMessage<PreloadedUserSettings_GuildFolders>): PreloadedUserSettings_GuildFolders {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.folders = [];
        message.guildPositions = [];
        if (value !== undefined)
            reflectionMergePartial<PreloadedUserSettings_GuildFolders>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PreloadedUserSettings_GuildFolders): PreloadedUserSettings_GuildFolders {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* repeated discord_protos.discord_users.v1.PreloadedUserSettings.GuildFolder folders */ 1:
                    message.folders.push(PreloadedUserSettings_GuildFolder.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* repeated fixed64 guild_positions */ 2:
                    if (wireType === WireType.LengthDelimited)
                        for (let e = reader.int32() + reader.pos; reader.pos < e;)
                            message.guildPositions.push(reader.fixed64().toBigInt());
                    else
                        message.guildPositions.push(reader.fixed64().toBigInt());
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PreloadedUserSettings_GuildFolders, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* repeated discord_protos.discord_users.v1.PreloadedUserSettings.GuildFolder folders = 1; */
        for (let i = 0; i < message.folders.length; i++)
            PreloadedUserSettings_GuildFolder.internalBinaryWrite(message.folders[i], writer.tag(1, WireType.LengthDelimited).fork(), options).join();
        /* repeated fixed64 guild_positions = 2; */
        if (message.guildPositions.length) {
            writer.tag(2, WireType.LengthDelimited).fork();
            for (let i = 0; i < message.guildPositions.length; i++)
                writer.fixed64(message.guildPositions[i]);
            writer.join();
        }
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message discord_protos.discord_users.v1.PreloadedUserSettings.GuildFolders
 */
export const PreloadedUserSettings_GuildFolders = new PreloadedUserSettings_GuildFolders$Type();